[#53] feat: 감정 일기 모듈 작성
